Verify google register (#83)
* fixed filters and sidebar to english

* passed things to english

* changed events to english

* changed seed

* fix spanish words

* addedd location toggle to edit event form

* fix: appears image in event progess

* removes id in friends when creating a group

* added google verification to not send email for verification

* added post searching in the homepage

---------

+++ backend/src/register/register.service.ts
@@ -35,24 +35,27 @@ export class RegisterService {
 
     const user = await this.userRepositoryService.createUser(registerDto);
 
-    const token = await this.jwtService.signAsync(
-      { sub: user.id, email: registerDto.email },
-      { expiresIn: '24h' },
-    );
+    if (!registerDto.isGoogleUser) {
+      const token = await this.jwtService.signAsync(
+        { sub: user.id, email: registerDto.email },
+        { expiresIn: '24h' },
+      );
 
-    try {
-      await this.mailService.sendVerificationEmail(
-        registerDto.email,
-        user.username,
-        token,
-      );
-    } catch (error) {
-      console.error('Failed to send verification email:', error);
-      await this.userRepositoryService.removeById(user.id);
-      throw new BadRequestException(
-        'Could not send verification email. Please try again later.',
-      );
+      try {
+        await this.mailService.sendVerificationEmail(
+          registerDto.email,
+          user.username,
+          token,
+        );
+      } catch (error) {
+        console.error('Failed to send verification email:', error);
+        await this.userRepositoryService.removeById(user.id);
+        throw new BadRequestException(
+          'Could not send verification email. Please try again later.',
+        );
+      }
     }
+
     return user;
   }
